+++ packages/apollo-link-schema/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 1.0.2
 - handle synchronous result of `execute` **Fixes #351**
+- pass context from operation into execute
 
 ### 1.0.1
 - export options as named interface [TypeScript]